Add files via upload
Looks through the data base for a student with a certian enumber and adds it.  All i need to do is get the add student window to pop up if the user wishes to enter a new student.  Also I need a way of letting the user type in an enumber to the window(there is no messagebox for the input of strings).